+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Dragonmark no longer add spell to spell list/free at will.
 - Mark of...  races no longer add too many extra spells to spell list/at will.
 - Reckless Attack tweaks
+- Wildshape did not always parse.
 
 
 # 6.4.19